feat: Internationalization with SSR
