+++ package.json
@@ -9,17 +9,21 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@formatjs/intl-localematcher": "^0.5.4",
     "@reduxjs/toolkit": "^2.2.4",
+    "negotiator": "^0.6.3",
     "next": "14.2.3",
     "next-images": "^1.8.5",
     "postcss-nested": "^6.0.1",
     "react": "^18",
     "react-dom": "^18",
     "react-redux": "^9.1.2",
+    "server-only": "^0.0.1",
     "tailwind-scrollbar": "^3.1.0"
   },
   "devDependencies": {
     "@playwright/test": "^1.44.0",
+    "@types/negotiator": "^0.6.3",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
