+++ src/app/[lang]/components/Banner/index.tsx
