+++ src/app/[lang]/components/MenuPanel/MenuCard/index.tsx
