+++ src/app/[lang]/components/MenuPanel/index.tsx
