+++ src/app/[lang]/globals.scss
